Use smlfut for raytracer.
This is slightly more inefficient than before, as smlfut does not
provide an option for copying a Futhark array into an existing array
(or slice).  That might be worth fixing in smlfut, but it will require
generating glue C code, as the FFI does not make it possible to do the
necessary pointer arithmetic.