Use newer smlfut with values_into to avoid a copy.
Unfortunately, this also required me to move from ForkJoin.alloc to
Int32Array.array.  I think that is fine for the raytracer, but it
could be a performance problem elsewhere, as it forces an early
initialisation.

Does MLton/MPL have a way to convert between polymorphic and
monomorphic array representations?  After all, they are identical at
runtime, as I understand it.